Serverless mode logs to console if file pipe not written

+++ newrelic-agent/src/main/java/com/newrelic/agent/transport/serverless/ServerlessWriterImpl.java
@@ -5,27 +5,38 @@ import com.newrelic.agent.logging.IAgentLogger;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
-import java.io.IOException;
 import java.util.logging.Level;
 
 public class ServerlessWriterImpl implements ServerlessWriter {
     private final IAgentLogger logger;
-    private final File pathFile;
+    private File pathFile;
+    private boolean fileCreated;
 
     public ServerlessWriterImpl(IAgentLogger logger, String filePath) {
         this.logger = logger;
-        pathFile = new File(filePath);
+        try {
+            pathFile = new File(filePath);
+            fileCreated = true;
+        } catch (Throwable t) {
+            fileCreated = false;
+            pathFile = null;
+        }
+
     }
 
     @Override
     public synchronized void write(String filePayload, String consolePayload) {
-        if (pathFile.exists()) {
+        boolean fileWritten = false;
+        if (fileCreated && pathFile.exists()) {
             try (BufferedWriter pipe = new BufferedWriter(new FileWriter(pathFile, false))) {
                 pipe.write(filePayload);
-            } catch (IOException ignored) {
-                logger.log(Level.FINEST, "Failed to write payload", ignored);
+                fileWritten = true;
+            } catch (Throwable ex) {
+                logger.log(Level.FINEST, "Failed to write payload, writing to console instead", ex);
             }
         }
-        System.out.println(consolePayload);
+        if (!fileWritten) {
+            System.out.println(consolePayload);
+        }
     }
 }
